Documents/workspace-spring-tool-suite-4-4.5.0.RELEASE/spring5app/src/main/java/es/iessoterohernandez/spring5app/business/entities/Product.java
package es.iessoterohernandez.spring5app.business.entities;

public class Product {
	
	private String description;
	private Double price;
	/**
	 * @return the description
	 */
	public String getDescription() {
		return description;
	}
	/**
	 * @param description the description to set
	 */
	public void setDescription(String description) {
		this.description = description;
	}
	/**
	 * @return the price
	 */
	public Double getPrice() {
		return price;
	}
	/**
	 * @param price the price to set
	 */
	public void setPrice(Double price) {
		this.price = price;
	}
	@Override
	public String toString() {
		return "Product [description=" + description + ", price=" + price + "]";
	}
	
	

}
