Documents/workspace-spring-tool-suite-4-4.5.0.RELEASE/spring5app/src/test/java/es/iessoterohernandez/spring5app/business/entities/ProductTest.java
package es.iessoterohernandez.spring5app.business.entities;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class ProductTest {
	
	private Product product;
	
	@BeforeEach
	public void setUp() throws Exception {
		product = new Product();
	}
	
	 @Test
    public void testSetAndGetDescription() {
        String testDescription = "aDescription";
        assertNull(product.getDescription());
        product.setDescription(testDescription);
        assertEquals(testDescription, product.getDescription());
    }

    @Test
    public void testSetAndGetPrice() {
        double testPrice = 100.00;
        assertEquals(0, 0, 0);    
        product.setPrice(testPrice);
        assertEquals(testPrice, product.getPrice(), 0);
	    }

}
